feat(navigation): separate route navigation
